+++ spec/unit/api_spec.rb
@@ -148,21 +148,44 @@ describe DogTrainer::API do
     end
   end
   describe '#generate_messages' do
-    it 'returns the appropriate message' do
-      expected = "{{#is_alert}}'mydesc' should be comp {{threshold}},"
-      expected += " but is {{value}}.{{/is_alert}}\n"
-      expected += "{{#is_recovery}}'mydesc' recovered  (current value "
-      expected += '{{value}} is comp threshold of {{threshold}}).'
-      expected += "{{/is_recovery}}\n(monitor and threshold configuration "
-      expected += 'for this alert is managed by my_repo_path) @my-notify-to'
-      msg, = subject.generate_messages('mydesc', 'comp')
-      expect(msg).to eq(expected)
+    context 'mon_type metric alert' do
+      it 'returns the appropriate message' do
+        expected = "{{#is_alert}}'mydesc' should be comp {{threshold}},"
+        expected += " but is {{value}}.{{/is_alert}}\n"
+        expected += "{{#is_recovery}}'mydesc' recovered  (current value "
+        expected += '{{value}} is comp threshold of {{threshold}}).'
+        expected += "{{/is_recovery}}\n(monitor and threshold configuration "
+        expected += 'for this alert is managed by my_repo_path) @my-notify-to'
+        msg, = subject.generate_messages('mydesc', 'comp', 'metric alert')
+        expect(msg).to eq(expected)
+      end
+      it 'returns the appropriate message' do
+        escalation = "'mydesc' is still in error state (current value {{value}}"
+        escalation += ' is comp threshold of {{threshold}})'
+        _, esc = subject.generate_messages('mydesc', 'comp', 'metric alert')
+        expect(esc).to eq(escalation)
+      end
     end
-    it 'returns the appropriate message' do
-      escalation = "'mydesc' is still in error state (current value {{value}}"
-      escalation += ' is comp threshold of {{threshold}})'
-      _, esc = subject.generate_messages('mydesc', 'comp')
-      expect(esc).to eq(escalation)
+    context 'mon_type service check' do
+      it 'returns the appropriate message' do
+        expected = "{{#is_alert}}'mydesc' is FAILING: {{check_message}}" \
+          "{{/is_alert}}\n"
+        expected += "{{#is_warning}}'mydesc' is WARNING: {{check_message}}" \
+          "{{/is_warning}}\n"
+        expected += "{{#is_recovery}}'mydesc' recovered: {{check_message}}" \
+          "{{/is_recovery}}\n"
+        expected += "{{#is_no_data}}'mydesc' is not reporting data" \
+          "{{/is_no_data}}\n"
+        expected += '(monitor and threshold configuration '
+        expected += 'for this alert is managed by my_repo_path) @my-notify-to'
+        msg, = subject.generate_messages('mydesc', 'comp', 'service check')
+        expect(msg).to eq(expected)
+      end
+      it 'returns the appropriate message' do
+        escalation = "'mydesc' is still in error state: {{check_message}}"
+        _, esc = subject.generate_messages('mydesc', 'comp', 'service check')
+        expect(esc).to eq(escalation)
+      end
     end
   end
   describe '#params_for_monitor' do
@@ -247,6 +270,68 @@ describe DogTrainer::API do
                mon_type: 'foo'
       )).to eq(expected)
     end
+    it 'sets renotify_interval if provided in options' do
+      expected = {
+        'name' => 'monname',
+        'type' => 'metric alert',
+        'query' => 'my_query',
+        'message' => 'my_msg',
+        'tags' => [],
+        'options' => {
+          'notify_audit' => false,
+          'locked' => false,
+          'timeout_h' => 0,
+          'silenced' => {},
+          'thresholds' => { 'critical' => 123.4 },
+          'require_full_window' => false,
+          'notify_no_data' => true,
+          'renotify_interval' => 120,
+          'no_data_timeframe' => 20
+        }
+      }
+      expect(subject.params_for_monitor(
+               'monname',
+               'my_msg',
+               'my_query',
+               123.4,
+               renotify_interval: 120
+      )).to eq(expected)
+    end
+    it 'passes through a threshold Hash' do
+      expected = {
+        'name' => 'monname',
+        'type' => 'foo',
+        'query' => 'my_query',
+        'message' => 'my_msg',
+        'tags' => [],
+        'options' => {
+          'notify_audit' => false,
+          'locked' => false,
+          'timeout_h' => 0,
+          'silenced' => {},
+          'thresholds' => {
+            'warning' => 50,
+            'critical' => 123.4,
+            'ok' => 20
+          },
+          'require_full_window' => false,
+          'notify_no_data' => true,
+          'renotify_interval' => 60,
+          'no_data_timeframe' => 20
+        }
+      }
+      expect(subject.params_for_monitor(
+               'monname',
+               'my_msg',
+               'my_query',
+               {
+                 'warning' => 50,
+                 'critical' => 123.4,
+                 'ok' => 20
+               },
+               mon_type: 'foo'
+      )).to eq(expected)
+    end
     it 'sets notify_no_data to false if provided in options' do
       expected = {
         'name' => 'monname',
@@ -289,7 +374,7 @@ describe DogTrainer::API do
           'thresholds' => { 'critical' => 123.4 },
           'require_full_window' => false,
           'notify_no_data' => false,
-          'renotify_interval' => 60,
+          'renotify_interval' => 123,
           'no_data_timeframe' => 20,
           'escalation_message' => 'myesc'
         }
@@ -301,7 +386,8 @@ describe DogTrainer::API do
                123.4,
                alert_no_data: false,
                mon_type: 'foo',
-               escalation_message: 'myesc'
+               escalation_message: 'myesc',
+               renotify_interval: 123
       )).to eq(expected)
     end
   end
@@ -321,11 +407,13 @@ describe DogTrainer::API do
         .and_return('12345')
 
       expect(dog).to_not receive(:update_monitor)
-      expect(subject).to receive(:generate_messages).once.with('mname', '>=')
+      expect(subject).to receive(:generate_messages).once
+        .with('mname', '>=', 'metric alert')
       expect(subject).to receive(:params_for_monitor).once
         .with('mname', 'msg', 'my_query', 123.4, escalation_message: 'esc',
                                                  alert_no_data: true,
-                                                 mon_type: 'metric alert')
+                                                 mon_type: 'metric alert',
+                                                 renotify_interval: 60)
       expect(subject).to receive(:get_existing_monitor_by_name).once
         .with('mname')
       expect(subject).to receive(:create_monitor).once
@@ -335,6 +423,41 @@ describe DogTrainer::API do
                'mname',
                'my_query',
                123.4,
+               '>='
+      )).to eq('12345')
+    end
+    it 'handles a hash threshold' do
+      params = { 'foo' => 'bar', 'baz' => 'blam' }
+      dog = double(Dogapi::Client)
+      allow(dog).to receive(:update_monitor).with(any_args)
+      subject.instance_variable_set('@dog', dog)
+      allow(subject).to receive(:generate_messages).with(any_args)
+        .and_return(%w(msg esc))
+      allow(subject).to receive(:params_for_monitor).with(any_args)
+        .and_return(params)
+      allow(subject).to receive(:get_existing_monitor_by_name).with(any_args)
+        .and_return(nil)
+      allow(subject).to receive(:create_monitor).with(any_args)
+        .and_return('12345')
+
+      expect(dog).to_not receive(:update_monitor)
+      expect(subject).to receive(:generate_messages).once
+        .with('mname', '>=', 'metric alert')
+      expect(subject).to receive(:params_for_monitor).once
+        .with('mname', 'msg', 'my_query', { 'warning' => 5.3, 'ok' => 1 },
+              escalation_message: 'esc',
+              alert_no_data: true,
+              mon_type: 'metric alert',
+              renotify_interval: 60)
+      expect(subject).to receive(:get_existing_monitor_by_name).once
+        .with('mname')
+      expect(subject).to receive(:create_monitor).once
+        .with('mname', params)
+
+      expect(subject.upsert_monitor(
+               'mname',
+               'my_query',
+               { 'warning' => 5.3, 'ok' => 1 },
                '>='
       )).to eq('12345')
     end
@@ -354,11 +477,13 @@ describe DogTrainer::API do
         .and_return('12345')
 
       expect(dog).to_not receive(:update_monitor)
-      expect(subject).to receive(:generate_messages).once.with('mname', '>=')
+      expect(subject).to receive(:generate_messages).once
+        .with('mname', '>=', 'metric alert')
       expect(subject).to receive(:params_for_monitor).once
         .with('mname', 'msg', 'my_query', 123.4, escalation_message: 'esc',
                                                  alert_no_data: true,
-                                                 mon_type: 'metric alert')
+                                                 mon_type: 'metric alert',
+                                                 renotify_interval: 60)
       expect(subject).to receive(:get_existing_monitor_by_name).once
         .with('mname')
       expect(subject).to_not receive(:create_monitor)
@@ -398,11 +523,13 @@ describe DogTrainer::API do
 
       expect(dog).to receive(:update_monitor).once
         .with('monid', 'my_query', params)
-      expect(subject).to receive(:generate_messages).once.with('mname', '>=')
+      expect(subject).to receive(:generate_messages).once
+        .with('mname', '>=', 'metric alert')
       expect(subject).to receive(:params_for_monitor).once
         .with('mname', 'msg', 'my_query', 123.4, escalation_message: 'esc',
                                                  alert_no_data: true,
-                                                 mon_type: 'metric alert')
+                                                 mon_type: 'metric alert',
+                                                 renotify_interval: 60)
       expect(subject).to receive(:get_existing_monitor_by_name).once
         .with('mname')
       expect(subject).to_not receive(:create_monitor)
@@ -437,11 +564,13 @@ describe DogTrainer::API do
 
       expect(dog).to receive(:update_monitor).once
         .with('monid', 'my_query', params)
-      expect(subject).to receive(:generate_messages).once.with('mname', '>=')
+      expect(subject).to receive(:generate_messages).once
+        .with('mname', '>=', 'metric alert')
       expect(subject).to receive(:params_for_monitor).once
         .with('mname', 'msg', 'my_query', 123.4, escalation_message: 'esc',
                                                  alert_no_data: true,
-                                                 mon_type: 'metric alert')
+                                                 mon_type: 'metric alert',
+                                                 renotify_interval: 60)
       expect(subject).to receive(:get_existing_monitor_by_name).once
         .with('mname')
       expect(subject).to_not receive(:create_monitor)
@@ -484,11 +613,13 @@ describe DogTrainer::API do
 
       expect(dog).to receive(:update_monitor).once
         .with('monid', 'my_query', params)
-      expect(subject).to receive(:generate_messages).once.with('mname', '>=')
+      expect(subject).to receive(:generate_messages).once
+        .with('mname', '>=', 'metric alert')
       expect(subject).to receive(:params_for_monitor).once
         .with('mname', 'msg', 'my_query', 123.4, escalation_message: 'esc',
                                                  alert_no_data: true,
-                                                 mon_type: 'metric alert')
+                                                 mon_type: 'metric alert',
+                                                 renotify_interval: 60)
       expect(subject).to receive(:get_existing_monitor_by_name).once
         .with('mname')
       expect(subject).to_not receive(:create_monitor)
@@ -518,11 +649,13 @@ describe DogTrainer::API do
         .and_return('12345')
 
       expect(dog).to_not receive(:update_monitor)
-      expect(subject).to receive(:generate_messages).once.with('mname', '>=')
+      expect(subject).to receive(:generate_messages).once
+        .with('mname', '>=', 'metric alert')
       expect(subject).to receive(:params_for_monitor).once
         .with('mname', 'msg', 'my_query', 123.4, escalation_message: 'esc',
                                                  alert_no_data: false,
-                                                 mon_type: 'metric alert')
+                                                 mon_type: 'metric alert',
+                                                 renotify_interval: 60)
       expect(subject).to receive(:get_existing_monitor_by_name).once
         .with('mname')
       expect(subject).to_not receive(:create_monitor)
@@ -551,11 +684,13 @@ describe DogTrainer::API do
         .and_return('12345')
 
       expect(dog).to_not receive(:update_monitor)
-      expect(subject).to receive(:generate_messages).once.with('mname', '>=')
+      expect(subject).to receive(:generate_messages).once
+        .with('mname', '>=', 'foobar')
       expect(subject).to receive(:params_for_monitor).once
         .with('mname', 'msg', 'my_query', 123.4, escalation_message: 'esc',
                                                  alert_no_data: true,
-                                                 mon_type: 'foobar')
+                                                 mon_type: 'foobar',
+                                                 renotify_interval: 60)
       expect(subject).to receive(:get_existing_monitor_by_name).once
         .with('mname')
       expect(subject).to_not receive(:create_monitor)
@@ -566,6 +701,221 @@ describe DogTrainer::API do
                123.4,
                '>=',
                mon_type: 'foobar'
+      )).to eq('monid')
+    end
+    it 'handles sparse options, with only renotify_interval' do
+      params = { 'foo' => 'bar', 'baz' => 'blam' }
+      existing = { 'foo' => 'bar', 'baz' => 'blam', 'id' => 'monid' }
+      dog = double(Dogapi::Client)
+      allow(dog).to receive(:update_monitor).with(any_args)
+      subject.instance_variable_set('@dog', dog)
+      allow(subject).to receive(:generate_messages).with(any_args)
+        .and_return(%w(msg esc))
+      allow(subject).to receive(:params_for_monitor).with(any_args)
+        .and_return(params)
+      allow(subject).to receive(:get_existing_monitor_by_name).with(any_args)
+        .and_return(existing)
+      allow(subject).to receive(:create_monitor).with(any_args)
+        .and_return('12345')
+
+      expect(dog).to_not receive(:update_monitor)
+      expect(subject).to receive(:generate_messages).once
+        .with('mname', '>=', 'metric alert')
+      expect(subject).to receive(:params_for_monitor).once
+        .with('mname', 'msg', 'my_query', 123.4, escalation_message: 'esc',
+                                                 alert_no_data: true,
+                                                 mon_type: 'metric alert',
+                                                 renotify_interval: 100)
+      expect(subject).to receive(:get_existing_monitor_by_name).once
+        .with('mname')
+      expect(subject).to_not receive(:create_monitor)
+
+      expect(subject.upsert_monitor(
+               'mname',
+               'my_query',
+               123.4,
+               '>=',
+               renotify_interval: 100
+      )).to eq('monid')
+    end
+    it 'handles sparse options, with only message' do
+      params = { 'foo' => 'bar', 'baz' => 'blam' }
+      existing = { 'foo' => 'bar', 'baz' => 'blam', 'id' => 'monid' }
+      dog = double(Dogapi::Client)
+      allow(dog).to receive(:update_monitor).with(any_args)
+      subject.instance_variable_set('@dog', dog)
+      allow(subject).to receive(:generate_messages).with(any_args)
+        .and_return(%w(msg esc))
+      allow(subject).to receive(:params_for_monitor).with(any_args)
+        .and_return(params)
+      allow(subject).to receive(:get_existing_monitor_by_name).with(any_args)
+        .and_return(existing)
+      allow(subject).to receive(:create_monitor).with(any_args)
+        .and_return('12345')
+
+      expect(dog).to_not receive(:update_monitor)
+      expect(subject).to receive(:generate_messages).once
+        .with('mname', '>=', 'metric alert')
+      expect(subject).to receive(:params_for_monitor).once
+        .with('mname', 'foo', 'my_query', 123.4, escalation_message: 'esc',
+                                                 alert_no_data: true,
+                                                 mon_type: 'metric alert',
+                                                 renotify_interval: 60)
+      expect(subject).to receive(:get_existing_monitor_by_name).once
+        .with('mname')
+      expect(subject).to_not receive(:create_monitor)
+
+      expect(subject.upsert_monitor(
+               'mname',
+               'my_query',
+               123.4,
+               '>=',
+               message: 'foo'
+      )).to eq('monid')
+    end
+    it 'handles sparse options, with only escalation_message' do
+      params = { 'foo' => 'bar', 'baz' => 'blam' }
+      existing = { 'foo' => 'bar', 'baz' => 'blam', 'id' => 'monid' }
+      dog = double(Dogapi::Client)
+      allow(dog).to receive(:update_monitor).with(any_args)
+      subject.instance_variable_set('@dog', dog)
+      allow(subject).to receive(:generate_messages).with(any_args)
+        .and_return(%w(msg esc))
+      allow(subject).to receive(:params_for_monitor).with(any_args)
+        .and_return(params)
+      allow(subject).to receive(:get_existing_monitor_by_name).with(any_args)
+        .and_return(existing)
+      allow(subject).to receive(:create_monitor).with(any_args)
+        .and_return('12345')
+
+      expect(dog).to_not receive(:update_monitor)
+      expect(subject).to receive(:generate_messages).once
+        .with('mname', '>=', 'metric alert')
+      expect(subject).to receive(:params_for_monitor).once
+        .with('mname', 'msg', 'my_query', 123.4, escalation_message: 'bar',
+                                                 alert_no_data: true,
+                                                 mon_type: 'metric alert',
+                                                 renotify_interval: 60)
+      expect(subject).to receive(:get_existing_monitor_by_name).once
+        .with('mname')
+      expect(subject).to_not receive(:create_monitor)
+
+      expect(subject.upsert_monitor(
+               'mname',
+               'my_query',
+               123.4,
+               '>=',
+               escalation_message: 'bar'
+      )).to eq('monid')
+    end
+    it 'handles sparse options, with only message and escalation_message' do
+      params = { 'foo' => 'bar', 'baz' => 'blam' }
+      existing = { 'foo' => 'bar', 'baz' => 'blam', 'id' => 'monid' }
+      dog = double(Dogapi::Client)
+      allow(dog).to receive(:update_monitor).with(any_args)
+      subject.instance_variable_set('@dog', dog)
+      allow(subject).to receive(:generate_messages).with(any_args)
+        .and_return(%w(msg esc))
+      allow(subject).to receive(:params_for_monitor).with(any_args)
+        .and_return(params)
+      allow(subject).to receive(:get_existing_monitor_by_name).with(any_args)
+        .and_return(existing)
+      allow(subject).to receive(:create_monitor).with(any_args)
+        .and_return('12345')
+
+      expect(dog).to_not receive(:update_monitor)
+      expect(subject).to receive(:generate_messages).once
+        .with('mname', '>=', 'metric alert')
+      expect(subject).to receive(:params_for_monitor).once
+        .with('mname', 'foo', 'my_query', 123.4, escalation_message: 'bar',
+                                                 alert_no_data: true,
+                                                 mon_type: 'metric alert',
+                                                 renotify_interval: 60)
+      expect(subject).to receive(:get_existing_monitor_by_name).once
+        .with('mname')
+      expect(subject).to_not receive(:create_monitor)
+
+      expect(subject.upsert_monitor(
+               'mname',
+               'my_query',
+               123.4,
+               '>=',
+               message: 'foo',
+               escalation_message: 'bar'
+      )).to eq('monid')
+    end
+    it 'handles sparse options, with only escalation_message set to nil' do
+      params = { 'foo' => 'bar', 'baz' => 'blam' }
+      existing = { 'foo' => 'bar', 'baz' => 'blam', 'id' => 'monid' }
+      dog = double(Dogapi::Client)
+      allow(dog).to receive(:update_monitor).with(any_args)
+      subject.instance_variable_set('@dog', dog)
+      allow(subject).to receive(:generate_messages).with(any_args)
+        .and_return(%w(msg esc))
+      allow(subject).to receive(:params_for_monitor).with(any_args)
+        .and_return(params)
+      allow(subject).to receive(:get_existing_monitor_by_name).with(any_args)
+        .and_return(existing)
+      allow(subject).to receive(:create_monitor).with(any_args)
+        .and_return('12345')
+
+      expect(dog).to_not receive(:update_monitor)
+      expect(subject).to receive(:generate_messages).once
+        .with('mname', '>=', 'metric alert')
+      expect(subject).to receive(:params_for_monitor).once
+        .with('mname', 'msg', 'my_query', 123.4, alert_no_data: true,
+                                                 mon_type: 'metric alert',
+                                                 renotify_interval: 60,
+                                                 escalation_message: nil)
+      expect(subject).to receive(:get_existing_monitor_by_name).once
+        .with('mname')
+      expect(subject).to_not receive(:create_monitor)
+
+      expect(subject.upsert_monitor(
+               'mname',
+               'my_query',
+               123.4,
+               '>=',
+               escalation_message: nil
+      )).to eq('monid')
+    end
+    it 'handles all options' do
+      params = { 'foo' => 'bar', 'baz' => 'blam' }
+      existing = { 'foo' => 'bar', 'baz' => 'blam', 'id' => 'monid' }
+      dog = double(Dogapi::Client)
+      allow(dog).to receive(:update_monitor).with(any_args)
+      subject.instance_variable_set('@dog', dog)
+      allow(subject).to receive(:generate_messages).with(any_args)
+        .and_return(%w(msg esc))
+      allow(subject).to receive(:params_for_monitor).with(any_args)
+        .and_return(params)
+      allow(subject).to receive(:get_existing_monitor_by_name).with(any_args)
+        .and_return(existing)
+      allow(subject).to receive(:create_monitor).with(any_args)
+        .and_return('12345')
+
+      expect(dog).to_not receive(:update_monitor)
+      expect(subject).to receive(:generate_messages).once
+        .with('mname', '>=', 'service check')
+      expect(subject).to receive(:params_for_monitor).once
+        .with('mname', 'foo', 'my_query', 123.4, escalation_message: 'bar',
+                                                 alert_no_data: false,
+                                                 mon_type: 'service check',
+                                                 renotify_interval: 10)
+      expect(subject).to receive(:get_existing_monitor_by_name).once
+        .with('mname')
+      expect(subject).to_not receive(:create_monitor)
+
+      expect(subject.upsert_monitor(
+               'mname',
+               'my_query',
+               123.4,
+               '>=',
+               renotify_interval: 10,
+               alert_no_data: false,
+               mon_type: 'service check',
+               message: 'foo',
+               escalation_message: 'bar'
       )).to eq('monid')
     end
   end